src/python/requirements.txt
syncedlyrics
spotipy
Flask
python-dotenv